# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Accounting API",
        "description": "Personal Accounting API empowers authorized users to seamlessly manage their personal finances by recording income and expenses.\n                  This API allows users to create, retrieve, and analyze financial transactions, providing a comprehensive solution for personal accounting.\n                  With secure authentication, intuitive endpoints, and insightful data summaries, users can easily track their financial activities, view transaction lists, and obtain cumulative amounts for both income and expenses.\n                  Simplify personal financial management with the capabilities of the Personal Accounting API.",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle an incoming authentication request",
                "operationId": "41355220271acb6e39d12f7db8e15749",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Returns User Model And Access Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "array",
                                    "title": "data",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "jK4vOOBBGy1tM0NOQWTY745xdPddS4IyvHRVNbDU34093fa7"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "18dac6b942e0d84b959ff170c665c6c2",
                "responses": {
                    "204": {
                        "description": "Destroy session",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User Model",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer \\<token\\>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}